Added sign in option to the 'View it' section
An alert with a 'Sign in' button is now shown in the 'View it' section. CSS is used to hide that alert if the user is logged in and to hide the alert in the 'Get it' section if there is already one in the 'View it' section.

Added some text via CSS to the 'Auto fill' buttons in the Document Delivery form to make it a bit clearer that they will use the DOI or PMID to auto-fill other sections.