Content for Database home page
We now show a notice on the Database search home page to advise that the search does not look within databases - it only searches for databases.

Text within the Database search home page's guided tour has also been modified with a similar message.

Added some styling for the 'Related reading' section that is shown for some resources.