Added content for the 'Browse' home page
The 'Browse' home page will now show some text saying what that search can be used for. (The functionality to do this was added to the implementation for the notice on the 'Databases' search, which is now a bit more general rather than just being done on the DB search.)

There was also a minor style change for the 'Browse' search page, which positions the pagination buttons closer to the list of items for larger screen widths.